refactor(agent-server): split OpenAI gateway service and models
