+++ tests/cross/test_remote_conversation_live_server.py
@@ -604,7 +604,7 @@ def test_openai_chat_completions_gateway_over_real_server(
     )
 
     with patch(
-        "openhands.agent_server.openai_router.LLMProfileStore",
+        "openhands.agent_server.openai_service.LLMProfileStore",
         lambda: LLMProfileStore(base_dir=profiles_dir),
     ):
         with live_server_env(tmp_path, monkeypatch) as env:
